moved to native_scanner_card.dart
