+++ lib/ui/cards/native_scanner/scanner_view.dart
